Read_description_because_i_added_a_lot
Intro scene was made but not finalized.
Buttons for start the tower defense game, Material gathering and upgrades (Didn't now what to call it, but it's meant to use the materials to aquire "perks" like when you send people scouting, they have a higher chance of surviving).
Some new adjustments to the dialogue system (now they can read from text files, so there is no need to code the dialogue).
The upgrade button is not functinal yet.
The scouting button sorta works, but  still needs some work